+++ src/Kit/General/Button/types.d.ts
@@ -2,7 +2,7 @@ import {KitHTMLAttributes} from '../../../types';
 import {CSSProperties, ForwardRefExoticComponent, RefAttributes} from 'react';
 import {IKitButtonPropetiesCssTokens} from '@theme/types/components/General/Button';
 
-export type KitButtonType = 'primary' | 'secondary' | 'tertiary' | 'text' | 'segmented';
+export type KitButtonType = 'primary' | 'secondary' | 'tertiary' | 'text' | 'segmented' | 'link' | 'redirect';
 
 type KitButtonHtmlType = 'submit' | 'button' | 'reset';
 
@@ -22,6 +22,8 @@ export type loadingConfig = {
 export interface IKitButton extends KitHTMLAttributes<HTMLDivElement> {
     type?: KitButtonType;
     htmlType?: KitButtonHtmlType;
+    href?: string;
+    target?: string;
     form?: string | undefined;
     iconSize?: KitButtonIconSize;
     disableSecureClick?: boolean;