+++ stories/02-Design System/01-General/Button/examples/index.tsx
@@ -7,6 +7,7 @@ import Danger from './danger';
 import SegmentedActive from './segmentedActive';
 import Colors from './segmentedColors';
 import Compact from './compact';
+import LinkRedirect from './redirect';
 
 import TypeSource from './type?raw';
 import IconSource from './icon?raw';
@@ -17,6 +18,7 @@ import DangerSource from './danger?raw';
 import SegmentedActiveSource from './segmentedActive?raw';
 import ColorsSource from './segmentedColors?raw';
 import CompactSource from './compact?raw';
+import LinkRedirectSource from './redirect?raw';
 
 export const Sources = {
     Type: TypeSource,
@@ -27,7 +29,8 @@ export const Sources = {
     Danger: DangerSource,
     SegmentedActive: SegmentedActiveSource,
     Colors: ColorsSource,
-    Compact: CompactSource
+    Compact: CompactSource,
+    LinkRedirect: LinkRedirectSource
 };
 
 export default {
@@ -39,5 +42,6 @@ export default {
     Danger,
     SegmentedActive,
     Colors,
-    Compact
+    Compact,
+    LinkRedirect
 };