+++ stories/02-Design System/01-General/Button/examples/loading.tsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import {KitButton, KitSpace} from '@kit/index';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import {faPowerOff} from '@fortawesome/free-solid-svg-icons';
+import {faMagnifyingGlass, faPowerOff, faWineBottle} from '@fortawesome/free-solid-svg-icons';
 
 const App = () => {
     const [loadings, setLoadings] = useState<boolean[]>([]);
@@ -33,6 +33,12 @@ const App = () => {
                     <KitButton type="text" loading>
                         Loading
                     </KitButton>
+                    <KitButton type="link" loading icon={<FontAwesomeIcon icon={faMagnifyingGlass} />}>
+                        Loading
+                    </KitButton>
+                    <KitButton type="redirect" loading icon={<FontAwesomeIcon icon={faWineBottle} />}>
+                        Loading
+                    </KitButton>
                     <KitButton type="primary" icon={<FontAwesomeIcon icon={faPowerOff} />} loading />
                     <KitButton type="segmented" icon={<FontAwesomeIcon icon={faPowerOff} />} loading />
                 </KitSpace>
@@ -51,16 +57,32 @@ const App = () => {
                         Click me!
                     </KitButton>
                     <KitButton
-                        type="primary"
-                        icon={<FontAwesomeIcon icon={faPowerOff} />}
+                        type="link"
+                        icon={<FontAwesomeIcon icon={faMagnifyingGlass} />}
                         loading={loadings[3]}
                         onClick={() => enterLoading(3)}
+                    >
+                        Click me!
+                    </KitButton>
+                    <KitButton
+                        type="redirect"
+                        icon={<FontAwesomeIcon icon={faWineBottle} />}
+                        loading={loadings[4]}
+                        onClick={() => enterLoading(4)}
+                    >
+                        Click me!
+                    </KitButton>
+                    <KitButton
+                        type="primary"
+                        icon={<FontAwesomeIcon icon={faPowerOff} />}
+                        loading={loadings[5]}
+                        onClick={() => enterLoading(5)}
                     />
                     <KitButton
                         type="segmented"
                         icon={<FontAwesomeIcon icon={faPowerOff} />}
-                        loading={loadings[4]}
-                        onClick={() => enterLoading(4)}
+                        loading={loadings[6]}
+                        onClick={() => enterLoading(6)}
                     />
                 </KitSpace>
             </KitSpace>