+++ CHANGELOG.md
@@ -37,6 +37,8 @@
 -   **Menu (Deprecated):**
     -   Deprecate Menu component
     -   Move Menu documentation to deprecated folder
+-   **Button**
+    - Add `link` and `redirect` Button types (@evoiron)
 
 ## [8.0.0](https://www.npmjs.com/package/aristid-ds/v/8.0.0) (2024-07-25)